Move getSync worker helper
