+++ package.json
@@ -19,6 +19,7 @@
   },
   "files": [
     "index.js",
+    "get-sync.js",
     "index.d.ts",
     "CHANGELOG.md",
     "UPGRADING.md"
